# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SalonSync API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: staffId
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithDetails"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
  /appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment (enforces cancellation fee policy)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelAppointmentBody"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelAppointmentResult"
  /availability:
    get:
      operationId: listAvailability
      tags: [availability]
      summary: List availability slots for staff
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
    post:
      operationId: createAvailability
      tags: [availability]
      summary: Create availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilityBody"
      responses:
        "201":
          description: Created availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
  /availability/{id}:
    delete:
      operationId: deleteAvailability
      tags: [availability]
      summary: Delete availability slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /ai/risk-score:
    post:
      operationId: scoreAppointmentRisk
      tags: [ai]
      summary: Score cancellation risk for an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskScoreBody"
      responses:
        "200":
          description: Risk score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScoreResult"
  /ai/sentiment:
    post:
      operationId: analyzeReviewSentiment
      tags: [ai]
      summary: Analyze sentiment of a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SentimentBody"
      responses:
        "200":
          description: Sentiment analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentimentResult"
  /ai/chat:
    post:
      operationId: aiReceptionistChat
      tags: [ai]
      summary: Chat with AI receptionist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResult"
components:
...
  schemas:
...
    SalonUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        locationId:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        specialties:
          type: array
          items:
            type: string
          nullable: true
        isActive:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, createdAt]
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [STANDARD, HIGH_VALUE]
          nullable: true
        basePrice:
          type: number
        durationMinutes:
          type: integer
        depositPercent:
          type: number
          nullable: true
        requiresFullPrepay:
          type: boolean
          nullable: true
        isActive:
          type: boolean
          nullable: true
        locationId:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, basePrice, durationMinutes, locationId, createdAt]
...
    AppointmentWithDetails:
      type: object
      properties:
        id:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
        notes:
          type: string
          nullable: true
        isHighValue:
          type: boolean
          nullable: true
        riskScore:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          nullable: true
        riskFactors:
          type: array
          items:
            type: string
          nullable: true
        totalPrice:
          type: number
        depositAmount:
          type: number
          nullable: true
        paymentStatus:
          type: string
          nullable: true
        cancelFeeCharged:
          type: number
          nullable: true
        cancelReason:
          type: string
          nullable: true
        locationId:
          type: string
        staffId:
          type: string
        clientId:
          type: string
        staff:
          $ref: "#/components/schemas/SalonUser"
          nullable: true
        client:
          $ref: "#/components/schemas/SalonUser"
          nullable: true
        services:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentServiceItem"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, startTime, endTime, status, totalPrice, locationId, staffId, clientId, createdAt]
    AppointmentServiceItem:
      type: object
      properties:
        id:
          type: string
        serviceId:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
      required: [id, serviceId, price, durationMinutes]
    CreateAppointmentBody:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        locationId:
          type: string
        staffId:
          type: string
        clientId:
          type: string
        serviceIds:
          type: array
          items:
            type: string
        notes:
          type: string
      required: [startTime, locationId, staffId, clientId, serviceIds]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, NO_SHOW]
        notes:
          type: string
        riskScore:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        riskFactors:
          type: array
          items:
            type: string
    CancelAppointmentBody:
      type: object
      properties:
        reason:
          type: string
        cancelledBy:
          type: string
      required: [cancelledBy]
    CancelAppointmentResult:
      type: object
      properties:
        appointment:
          $ref: "#/components/schemas/AppointmentWithDetails"
        feeCharged:
          type: number
          nullable: true
        message:
          type: string
      required: [appointment, message]
    Availability:
      type: object
      properties:
        id:
          type: string
        dayOfWeek:
          type: integer
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        isBlocked:
          type: boolean
          nullable: true
        blockDate:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
          nullable: true
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, createdAt]
    CreateAvailabilityBody:
      type: object
      properties:
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isBlocked:
          type: boolean
        blockDate:
          type: string
          format: date-time
        note:
          type: string
        userId:
          type: string
      required: [userId]
...
    RiskScoreBody:
      type: object
      properties:
        appointmentId:
          type: string
        clientId:
          type: string
        serviceIds:
          type: array
          items:
            type: string
        startTime:
          type: string
          format: date-time
        clientHistory:
          type: object
          properties:
            totalAppointments:
              type: integer
            cancelledCount:
              type: integer
            noShowCount:
              type: integer
      required: [appointmentId, clientId, serviceIds, startTime]
    RiskScoreResult:
      type: object
      properties:
        riskScore:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        riskFactors:
          type: array
          items:
            type: string
      required: [riskScore, riskFactors]
    SentimentBody:
      type: object
      properties:
        reviewId:
          type: string
        comment:
          type: string
        rating:
          type: integer
      required: [reviewId, rating]
    SentimentResult:
      type: object
      properties:
        sentimentScore:
          type: number
        sentimentTags:
          type: array
          items:
            type: string
      required: [sentimentScore, sentimentTags]
    ChatBody:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required: [role, content]
        locationId:
          type: string
      required: [messages, locationId]
    ChatResult:
      type: object
      properties:
        response:
          type: string
      required: [response]