# lib/api-zod/src/generated/types/cancelAppointmentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonSync API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentWithDetails } from "./appointmentWithDetails";

export interface CancelAppointmentResult {
  appointment: AppointmentWithDetails;
  feeCharged?: number | null;
  message: string;
}