# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonSync API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from "./serviceCategory";

export interface Service {
  id: string;
  name: string;
  description?: string | null;
  category?: ServiceCategory;
  basePrice: number;
  durationMinutes: number;
  depositPercent?: number | null;
  requiresFullPrepay?: boolean | null;
  isActive?: boolean | null;
  locationId: string;
  createdAt: Date;
}